src/tweetscheduler/requirements.txt
certifi==2019.6.16
chardet==3.0.4
Django==2.2.3
djangorestframework==3.10.1
future==0.17.1
idna==2.8
mysql-connector-python==8.0.16
mysqlclient==1.4.2.post1
oauthlib==3.0.2
protobuf==3.9.0
PySocks==1.7.0
python-twitter==3.5
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.3.0
tweepy==3.8.0
urllib3==1.25.3
